requirements.txt
xlrd==1.2.0
yfinance==0.1.54
pandas==1.0.5
streamlit==1.1.0
lxml==4.5.2
numpy==1.19.5
altair==4.1.0
python_dateutil==2.8.2
